Add files via upload
A background color changer is a feature that allows users to change the background color of a website or web page dynamically. It's a popular and simple way to enhance user experience and add a little bit of interactivity to a web page.

The functionality of a background color changer is typically implemented using JavaScript, which allows the website to dynamically modify the CSS styles of the page in response to user actions such as clicking a button or selecting a color from a color palette.

A background color changer can be used for a variety of purposes. It can help to highlight specific content on a page, improve readability, create a specific mood or atmosphere, or simply add some fun and interactivity to a website.

Web developers can easily add a background color changer to their websites using JavaScript and CSS. They can customize the feature to suit their specific needs by selecting a color palette, defining the behavior of the color changer, and integrating it with other interactive elements on the page.

Overall, a background color changer is a simple yet effective way to add some life and interactivity to a website. With its easy implementation and customization, it's a great tool for web developers who want to enhance user experience on their websites.